X2-4697 - Submit date range on submit of relative date range picker.
We now maintain an "internal" state for the date picker so we can decide when and when not to pass it on to the onChange callback.
This is done so that the seller app does not have to maintain a value
for the date picker and a value for the input field.