Do not checkout & run workflow if user is not in Xola
This is so that Xola team members can run the workflow from
`ataylorme/eslint-annotate-action@2.2.0` that writes to checks and
pull requests. This needs `pull_request_target` because for public repos
Github does not give write access to secrets.GITHUB_TOKEN and for some
reason giving a PT to the `eslint-annotate-action` workflow does not
work.

If we changed the default to `pull_request_target` this will allow
any user to run code in Github actions potentially giving them write
access